Add folder for skills helper funcs, refactor lockpick skill.
